# Conflicts:
#	indexing-service/src/main/java/org/apache/druid/indexing/overlord/ForkingTaskRunner.java

+++ sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/IPv4AddressParseOperatorConversion.java
@@ -39,7 +39,7 @@ public class IPv4AddressParseOperatorConversion extends DirectOperatorConversion
               OperandTypes.family(SqlTypeFamily.STRING),
               OperandTypes.family(SqlTypeFamily.INTEGER)
           ))
-      .returnTypeNullable(SqlTypeName.INTEGER)
+      .returnTypeNullable(SqlTypeName.BIGINT)
       .functionCategory(SqlFunctionCategory.USER_DEFINED_FUNCTION)
       .build();
